requirements.txt
python-telegram-bot==20.7
yt-dlp==2023.12.30
firebase-admin==6.5.0
